requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Migrate==4.0.7
bleach==6.1.0
Markdown==3.6
PyMySQL==1.1.1
SQLAlchemy==2.0.29
MarkupSafe==2.1.5
WTForms==3.1.2
pytest
pytest-flask 